feat: implement Motorsport library modules for class, driver, and duration management
