+++ cli/motorsport/src/duration.rs
@@ -1,0 +1,70 @@
+pub type Duration = u32;
+
+/// Duration文字列をミリ秒に変換
+/// 例: "1:35.365" -> 95365, "23.155" -> 23155
+pub fn from_string(s: &str) -> Option<Duration> {
+    let parts: Vec<&str> = s.split(':').collect();
+
+    match parts.as_slice() {
+        // "hh:mm:ss.ms" 形式
+        [h, m, s] => {
+            let hours = h.parse::<u32>().ok()?;
+            let minutes = m.parse::<u32>().ok()?;
+            let seconds = s.parse::<f64>().ok()?;
+            Some(hours * 3600000 + minutes * 60000 + (seconds * 1000.0) as u32)
+        }
+        // "mm:ss.ms" 形式
+        [m, s] => {
+            let minutes = m.parse::<u32>().ok()?;
+            let seconds = s.parse::<f64>().ok()?;
+            Some(minutes * 60000 + (seconds * 1000.0) as u32)
+        }
+        // "ss.ms" 形式
+        [s] => {
+            let seconds = s.parse::<f64>().ok()?;
+            Some((seconds * 1000.0) as u32)
+        }
+        _ => None,
+    }
+}
+
+/// ミリ秒をDuration文字列に変換
+pub fn to_string(ms: Duration) -> String {
+    if ms >= 3600000 {
+        // 1時間以上
+        let hours = ms / 3600000;
+        let minutes = (ms % 3600000) / 60000;
+        let seconds = ((ms % 60000) as f64) / 1000.0;
+        format!("{}:{:02}:{:06.3}", hours, minutes, seconds)
+    } else if ms >= 60000 {
+        // 1分以上
+        let minutes = ms / 60000;
+        let seconds = ((ms % 60000) as f64) / 1000.0;
+        format!("{}:{:06.3}", minutes, seconds)
+    } else {
+        // 1分未満
+        let seconds = (ms as f64) / 1000.0;
+        format!("{:.3}", seconds)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{from_string, to_string};
+
+    #[test]
+    fn test_duration_from_string() {
+        assert_eq!(from_string("1:35.365"), Some(95365));
+        assert_eq!(from_string("23.155"), Some(23155));
+        assert_eq!(from_string("0:29.928"), Some(29928));
+        assert_eq!(from_string("7:06:54.321"), Some(25614321));
+    }
+
+    #[test]
+    fn test_duration_to_string() {
+        assert_eq!(to_string(95365), "1:35.365");
+        assert_eq!(to_string(23155), "23.155");
+        assert_eq!(to_string(29928), "29.928");
+        assert_eq!(to_string(25614321), "7:06:54.321");
+    }
+}
